fix : filename and dynamic day transformation
